backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.6.15
click==8.1.8
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.15
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.5
pydantic-settings==2.8.1
pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
starlette==0.44.0
typer==0.16.0
typing-extensions==4.13.2
ujson==5.10.0
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
